requirements.txt
asgiref
certifi
cffi
charset-normalizer
crispy-bootstrap5
cryptography
defusedxml
Django
django-allauth
django-crispy-forms
django-debug-toolbar
gunicorn
idna
oauthlib
packaging
psycopg
psycopg-binary
pycparser
PyJWT
python-openid
requests
requests-oauthlib
sqlparse
typing_extensions
urllib3
whitenoise
Django-Verify-Email
01OS